+++ pkg/web/api_01.go
@@ -4,19 +4,36 @@ func api1() {
 
 	a := e.Group("/api/v1")
 
-	// USER
+	// User
 
 	u := a.Group("/user")
 
 	u.POST("/", userCreate)
 	u.POST("/login/", userLogin)
 
-	p := u.Group("/profile")
+	pr := u.Group("/profile")
 
-	useJWT(p)
+	useJWT(pr)
 
-	p.GET("/", userRead)
-	p.PUT("/", userUpdate)
-	p.DELETE("/", userDelete)
+	pr.GET("/", userRead)
+	pr.PUT("/", userUpdate)
+	pr.DELETE("/", userDelete)
+
+	// Library
+
+	l := a.Group("/library")
+
+	useJWT(l)
+
+	l.POST("/", libraryAdd)
+	l.GET("/", libraryView)
+
+	li := l.Group("/:audio")
+
+	li.GET("/", songView)
+	li.PUT("/", songUpdate)
+	li.DELETE("/", songDelete)
+
+	// Player
 
 }