+++ CHANGELOG.md
@@ -8,6 +8,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - Web Interface
 - Distribution code
+- Tags
+- Thumbnails
+
+## [0.3.0] - 2018-08-21
+
+### Added
+
+- 'init' command to generate config file.
+- Ability to import files through web API.
+- View all add tracks through web API.
+
+### Fixed
+
+- Login attempts no longer print.
+- Audio files are removed when related database entries are removed.
 
 ## [0.2.0] - 2018-08-21